+++ src/components/App/App.tsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import s from "./App.module.scss";
 import Rampa from "../../asset/rampa.svg";
 import Toggle from "../../asset/toggle.svg";
-import { useTheme } from "../../theme/useTheme";
+import { useTheme } from "../../styles/theme/useTheme";
 
 const currentWindow = window as any;
 const tg = currentWindow.Telegram.WebApp;
