--- src/store/constants.ts
@@ -1,1 +1,0 @@
-export const DIFFICULTY_LEVELS = ["Easy", "Medium", "Hard", "Impossible"];
