+++ src/styles/theme/ThemeContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import { ThemeContextProps } from "./types";
+
+export const ThemeContext = createContext<ThemeContextProps>({});
