+++ src/styles/theme/ThemeProvider.tsx
@@ -1,15 +1,13 @@
 import { FC, useMemo, useState } from "react";
-import { LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext } from "./ThemeContext";
+import { ThemeContext } from "./ThemeContext";
+import { Theme, ThemeProviderProps } from "./types";
+import { LOCAL_STORAGE_THEME_KEY } from "utils/constants";
 
-const defaultThene =
+const defaultTheme =
   (localStorage.getItem(LOCAL_STORAGE_THEME_KEY) as Theme) || Theme.LIGHT;
 
-interface Props {
-  children: React.ReactNode;
-}
-
-const ThemeProvider: FC<Props> = ({ children }) => {
-  const [theme, setTheme] = useState<Theme>(defaultThene);
+const ThemeProvider: FC<ThemeProviderProps> = ({ children }) => {
+  const [theme, setTheme] = useState<Theme>(defaultTheme);
 
   const defaultProps = useMemo(
     () => ({
