+++ src/styles/theme/useTheme.ts
@@ -1,10 +1,7 @@
 import { useContext } from "react";
-import { LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext } from "./ThemeContext";
-
-interface UseThemeResult {
-  toggleTheme: () => void;
-  theme: Theme;
-}
+import { ThemeContext } from "./ThemeContext";
+import { Theme, UseThemeResult } from "./types";
+import { LOCAL_STORAGE_THEME_KEY } from "utils/constants";
 
 export function useTheme(): UseThemeResult {
   const { theme, setTheme } = useContext(ThemeContext);
