+++ src/styles/vars.scss
@@ -1,13 +1,15 @@
 :root {
-  --color-black: rgb(226, 218, 218);
-  --color-aquamarine: red;
+  // dark-theme-start
   --bg-dark: #7298a4;
   --primary-dark-text: #e3e5d5;
   --secondary-dark-text: #d4d6c8;
   --btn-dark-color: #404571;
+  // dark-theme-end
 
+  // light-theme-start
   --bg-light: #ffffff;
   --primary-light-text: #1a1e1f;
   --secondary-light-text: #5c676b;
   --btn-light-color: #980227;
+  // light-theme-end
 }
